requirements.txt
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
exceptiongroup==1.1.1
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
outcome==1.2.0
packaging==23.1
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
python-whois==0.8.0
requests==2.30.0
selenium==4.9.1
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
zstandard==0.21.0
